requirements.txt
alembic==1.8.0
amqp==5.1.1
async-timeout==4.0.2
Babel==2.10.1
bcrypt==3.2.2
bidict==0.22.0
billiard==3.6.4.0
blinker==1.6.2
cachelib==0.9.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
croniter==1.3.7
cryptography==38.0.4
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.6
dnspython==2.2.1
email-validator==1.2.1
environs==9.5.0
fakeredis==1.10.1
filelock==3.12.0
Flask==2.3.2
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-CLI==0.4.0
Flask-Cors==3.0.10
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-redis==0.4.0
Flask-Session==0.4.0
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==3.0.3
Flask-SSLify==0.1.5
Flask-WTF==1.1.1
flower==1.2.0
gevent==21.12.0
gevent-websocket==0.10.1
google-auth==2.11.0
google-auth-oauthlib==0.5.2
greenlet==1.1.2
gunicorn==20.1.0
http-ece==1.1.0
humanize==4.4.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.16.0
mockredis==0.1.3.dev0
numpy==1.23.4
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pipenv==2023.4.29
platformdirs==3.5.0
prometheus-client==0.15.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.5
py-vapid==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.2
python-socketio==5.6.0
python-webnoti==0.4.1
pytz==2022.1
pywebpush==1.14.0
redis==4.3.3
redis-queue==0.5
requests==2.27.1
requests-oauthlib==1.3.1
rq==1.11.1
rq-scheduler==0.11.0
rsa==4.9
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.15
SQLAlchemy-Utils==0.41.1
tornado==6.2
typing_extensions==4.5.0
ua-parser==0.16.1
urllib3==1.26.9
uWSGI==2.0.20
validators==0.20.0
vine==5.0.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.3.4
wrapt==1.14.1
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
